Alec's first collections commit: warning, stupidly large.
This commit includes a complete refactoring of the collections system to use standard Drupal nodes instead of custom collection objects. Changes rely upon an assumed DB structure that will be established programmatically through a Features module (to be committed later). Data migration will be achieved through the codeincluded in the next commit. Notation has hopefully improved in the collection module; note that it retains Nelson's structure and should be more significantly refactored in the future (particularly the view-related components, which confuse presentation and logic greatly).